Rename validate/josev2 pkg to validator
