+++ go.mod
@@ -3,7 +3,6 @@ module github.com/auth0/go-jwt-middleware
 go 1.17
 
 require (
-	github.com/golang-jwt/jwt/v4 v4.1.0
 	github.com/google/go-cmp v0.5.6
 	github.com/pkg/errors v0.9.1
 	gopkg.in/square/go-jose.v2 v2.5.1
