+++ validator/examples/README.md
@@ -82,6 +82,6 @@ The token isn't valid: expected claims not validated: square/go-jose/jwt: valida
 ```
 
 ### JWKS
-For a JWKS example please see [examples/http-jwks-example/README.md](../../../examples/http-jwks-example/README.md).
+For a JWKS example please see [examples/http-jwks-example/README.md](../../examples/http-jwks-example/README.md).
 
 Take a look through the example code and things will make a lot more sense.
